api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=12.0
sqlalchemy>=2.0.25
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
openai>=1.57.0
anthropic>=0.40.0
google-generativeai>=0.8.0
parsec-llm>=0.1.0
